Added initial version of microORM

+++ src/MyFramework/Contracts/IAuthEntity.cs
@@ -2,6 +2,6 @@
 
 public interface IAuthEntity
 { 
-    public long UserId { get; init; }
+    public int UserId { get; init; }
     public string? Role { get; init; }
 }
